requirements.txt
Flask==2.2.3
gunicorn==21.2.0
keras==2.13.1
numpy==1.23.5
tensorflow==2.13.0
Werkzeug==2.2.3
pillow==10.2.0
boto3==1.34.44